Optimize FAQ for copying and pasting. (#223)
It seems that these are not read as often as I would hope in their
present location. Changing them to a code block means it will be
slightly more straightforward for me to move them to where they will be
read.

Also, added an FAQ on not using the course materials.